Mouse events now work and affect the border layers with the animation on.
Just wanted to get the feature to work so something visible and testable
happens on the screen. Still not working:

- Easing functions and modes
- Controls where IsAnimated is set to false
- Some other smaller miscellaneous stuff

+++ BendUI.Controls/Drawing/BorderLayer.cs
@@ -155,7 +155,7 @@ namespace BendUI.Controls.Drawing
 		}
 
 		// When animating transitions, drawing tools will be refreshed every millisecond
-		public override void RefreshDrawingTools(int currentPercentage)
+		public override void RefreshDrawingTools(float currentPercentage)
 		{
 			if (currentPercentage == 0)
 			{
